efitools: Allow building an ESL from a list of hashes
At this moment hash-to-efi-sig-list only allows to add hash to an ESL
if the whole EFI application is presented, but this is technically
not necessary. We want to be able to use raw hashes, because
for PCIe devices using 3rd party drivers, we only know the hash,
the full EFI application or driver is not accessible to userspace.

This patch adds support to use raw hashes to hash-to-efi-sig-list
and ships the tool with efitools-utils.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>