balena-init-flasher-efi: enroll hashes from PCR2 during provisioning
At this moment, if anything is measured into PCR2 (OpROMs, UEFI drivers)
the device will provision, but not be able to boot the first time.
This is because the EFI applications measured in PCR2 need to verify
against `db`, which means they will be loaded during provisioning
(while secure boot is still off), but won't be loaded once secure boot
is enabled. This changes the value of PCR2 between provisioning
and first boot, which means the encryption keys won't be released.

This patch makes flasher read the hashes present in PCR2 and enroll
them into `db` during provisioning.

Change-type: minor
Signed-off-by: Michal Toman <michalt@balena.io>